Last Commit Before Friday
Ok so what has changed is the API calls, the event class (not only has this class changed, Ryan is going to change it again before tomorrow) and the userlist which im working on populating

Basically everything will have to change again becuase the table was changed again.

What you should do tomorrow is try to get the events displaying on the calendar. There are a few exampels of using the API to pull data. One is in the logindialog. The other is in the userlist. The final one is in browse. They all follow the same structure. You must write the request in the dbOperation in the API and then define the case in the API file (both in php). Then you have to do a few things in java. First perform the network request (just copy this code from the previous classes) and then inside there on PostExecute call a method that redraws or draws the calendar / whatever you are trying to create.